Actually cycle through GPU IDs in find_suitable_gpu
In the current version, if the first returned GPU does not have the
desired features, we are stuck in an endless loop, since we always
retry with the exact same ID.

This patch adds an increment to core_id, assuming thats how we are
supposed to cycle through the GPUs.

Tested on an IMX6Q6AVT10AD where it works with this patch applied.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>